some changes
refresh got some examples of how it should be refactored

discovered that in this version mana was not being reduced as it should

mana is not being regenerated also

mana is broken, basically. now it can even become negative.

will not fix this as I am working in a new Skill class that should change
some stuff, then we fix the mana problem.

mafagafo